requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-deepseek>=0.1.0
python-dotenv>=1.0.0
duckduckgo-search>=3.9.0
langchain-community>=0.0.1